[OCTOBER] Installed plugin "rainlab.blog" and updated the system and all plugins.

+++ plugins/rainlab/user/controllers/Users.php
@@ -1,5 +1,6 @@
 <?php namespace RainLab\User\Controllers;
 
+use Auth;
 use Lang;
 use Flash;
 use BackendMenu;
@@ -8,6 +9,7 @@ use Backend\Classes\Controller;
 use System\Classes\SettingsManager;
 use RainLab\User\Models\User;
 use RainLab\User\Models\UserGroup;
+use RainLab\User\Models\MailBlocker;
 use RainLab\User\Models\Settings as UserSettings;
 
 class Users extends Controller
@@ -31,6 +33,13 @@ class Users extends Controller
 
         BackendMenu::setContext('RainLab.User', 'user', 'users');
         SettingsManager::setContext('RainLab.User', 'settings');
+    }
+
+    public function index()
+    {
+        $this->addJs('/plugins/rainlab/user/assets/js/bulk-actions.js');
+
+        $this->asExtension('ListController')->index();
     }
 
     /**
@@ -73,10 +82,27 @@ class Users extends Controller
         }
     }
 
+    public function formAfterUpdate($model)
+    {
+        $blockMail = post('User[block_mail]', false);
+        if ($blockMail !== false) {
+            $blockMail ? MailBlocker::blockAll($model) : MailBlocker::unblockAll($model);
+        }
+    }
+
+    public function formExtendModel($model)
+    {
+        $model->block_mail = MailBlocker::isBlockAll($model);
+
+        $model->bindEvent('model.saveInternal', function() use ($model) {
+            unset($model->attributes['block_mail']);
+        });
+    }
+
     /**
      * Manually activate a user
      */
-    public function update_onActivate($recordId = null)
+    public function preview_onActivate($recordId = null)
     {
         $model = $this->formFindModelObject($recordId);
 
@@ -84,7 +110,23 @@ class Users extends Controller
 
         Flash::success(Lang::get('rainlab.user::lang.users.activated_success'));
 
-        if ($redirect = $this->makeRedirect('update', $model)) {
+        if ($redirect = $this->makeRedirect('update-close', $model)) {
+            return $redirect;
+        }
+    }
+
+    /**
+     * Manually unban a user
+     */
+    public function preview_onUnban($recordId = null)
+    {
+        $model = $this->formFindModelObject($recordId);
+
+        $model->unban();
+
+        Flash::success(Lang::get('rainlab.user::lang.users.unbanned_success'));
+
+        if ($redirect = $this->makeRedirect('update-close', $model)) {
             return $redirect;
         }
     }
@@ -106,23 +148,49 @@ class Users extends Controller
     }
 
     /**
-     * Deleted checked users
+     * Perform bulk action on selected users
      */
-    public function index_onDelete()
+    public function index_onBulkAction()
     {
-        if (($checkedIds = post('checked')) && is_array($checkedIds) && count($checkedIds)) {
+        if (
+            ($bulkAction = post('action')) &&
+            ($checkedIds = post('checked')) &&
+            is_array($checkedIds) &&
+            count($checkedIds)
+        ) {
 
             foreach ($checkedIds as $userId) {
-                if (!$user = User::find($userId)) {
+                if (!$user = User::withTrashed()->find($userId)) {
                     continue;
                 }
-                $user->delete();
+
+                switch ($bulkAction) {
+                    case 'delete':
+                        $user->forceDelete();
+                        break;
+
+                    case 'deactivate':
+                        $user->delete();
+                        break;
+
+                    case 'restore':
+                        $user->restore();
+                        break;
+
+                    case 'ban':
+                        $user->ban();
+                        break;
+
+                    case 'unban':
+                        $user->unban();
+                        break;
+                }
             }
 
-            Flash::success(Lang::get('rainlab.user::lang.users.delete_selected_success'));
+            Flash::success(Lang::get('rainlab.user::lang.users.'.$bulkAction.'_selected_success'));
         }
         else {
-            Flash::error(Lang::get('rainlab.user::lang.users.delete_selected_empty'));
+            Flash::error(Lang::get('rainlab.user::lang.users.'.$bulkAction.'_selected_empty'));
         }
 
         return $this->listRefresh();
